feat: TS version of babel
