+++ packages/goat-babel/package.json
@@ -2,9 +2,13 @@
   "name": "@the-goat/babel",
   "version": "1.12.1",
   "description": "",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "npm run clean && npm run compile",
+    "clean": "rimraf -rf ./dist",
+    "compile": "../../node_modules/.bin/tsc -p tsconfig.build.json",
+    "prepublishOnly": "npm run build",
+    "test": "npm run build"
   },
   "author": "Stef Spakman <goat@stefspakman.com>",
   "license": "MIT",
@@ -12,11 +16,15 @@
     "@babel/core": "^7.12.10",
     "@babel/plugin-transform-runtime": "^7.12.10",
     "@babel/runtime": "^7.12.5",
+    "@the-goat/goat": "*",
     "@rollup/plugin-node-resolve": "^11.0.1",
     "@rollup/plugin-virtual": "^2.0.3",
     "babel-preset-airbnb": "^5.0.0",
     "lodash": "^4.17.20",
     "rollup": "^2.35.1"
+  },
+  "devDependencies": {
+    "rimraf": "^2.6.2"
   },
   "gitHead": "cee2b439089c6f90cdf345d85bbf0ae4414988b3",
   "goat": {
